+++ src/quickmock.sh
@@ -7,6 +7,7 @@ SQM_Init()
     local src="$(dirname $(readlink -f $BASH_SOURCE))"
 
     . ${src}/support.sh
+    . ${src}/common.sh
     . ${src}/stub.sh
 }