+++ src/stub.sh
@@ -1,7 +1,5 @@
 #! /bin/bash
 # quickmock stubbing library
-# provides some global variables:
-#   - QUICKMOCK_STUBS: list of all existing stubs (should not be updated directly)
 
 # create a new stub
 # @param string $1 name of the command to get a stub of
@@ -21,32 +19,5 @@ QuickMock.newStub()
 
     eval "${name}() { ${body}; }"
 
-    QuickMock.trackStub "${name}"
-}
-
-# track a new stub
-# @param string $1 name of the stub to track
-QuickMock.trackStub()
-{
-    QUICKMOCK_STUBS="${QUICKMOCK_STUBS} $1"
-    QUICKMOCK_STUBS="$(QuickMock.support.normalizeSpaces ${QUICKMOCK_STUBS})"
-}
-
-# release one or many stubs
-# @param string[] $@ list of stub names to release
-QuickMock.releaseStubs()
-{
-    local stubs="$@"
-
-    if [ -z "${stubs}" ]; then
-        stubs="${QUICKMOCK_STUBS}"
-    fi
-
-    for stub in ${stubs[@]}; do
-        QUICKMOCK_STUBS="${QUICKMOCK_STUBS/${stub}/}"
-        unset "${stub}"
-        QuickMock.support.revertFunction "${stub}"
-    done
-
-    QUICKMOCK_STUBS="$(QuickMock.support.normalizeSpaces ${QUICKMOCK_STUBS})"
+    QuickMock.trackDouble "${name}"
 }